# doc-files/kypo-user-and-group-swagger-open-api.yaml
...
swagger: "2.0"
info:
  version: "Version: 1.1.77"
  title: "KYPO User and Group - API Reference"
host: "localhost:8084"
basePath: "/kypo-rest-user-and-group/api/v1"
...
schemes:
- "http"
- "https"
paths:
  /groups:
    get:
      tags:
      - "groups"
      summary: "Get groups."
      description: ""
      operationId: "getGroups"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Filtering on IDMGroup entity attributes"
        required: false
        schema:
          $ref: "#/definitions/Predicate"
      - name: "fields"
        in: "query"
        description: "Fields which should be returned in REST API response"
        required: false
        type: "string"
      - name: "page"
        in: "query"
        description: "Results page you want to retrieve (0..N)"
        required: false
        type: "int"
        x-example: "0"
      - name: "size"
        in: "query"
        description: "Number of records per page."
        required: false
        type: "int"
        x-example: "20"
      - name: "sort"
        in: "query"
        description: "Sorting criteria in the format: property(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        type: "array"
        items:
          type: "string"
          example: "asc"
        collectionFormat: "multi"
        x-example: "asc"
      responses:
        200:
          description: "All groups found."
          schema:
            $ref: "#/definitions/GroupRestResource"
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
    post:
      tags:
      - "groups"
      summary: "Create new group."
      description: ""
      operationId: "createNewGroup"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Group to be created."
        required: true
        schema:
          $ref: "#/definitions/NewGroupDTO"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/GroupDTO"
        201:
          description: "Given group created."
          schema:
            $ref: "#/definitions/GroupDTO"
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
    put:
      tags:
      - "groups"
      summary: "Update group."
      description: ""
      operationId: "updateGroup"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Group to be updated."
        required: true
        schema:
          $ref: "#/definitions/UpdateGroupDTO"
      responses:
        200:
          description: "Group updated."
        405:
          description: "Group is external and cannot be modified."
          schema:
            $ref: "#/definitions/ApiError"
        409:
          description: "Name of the main group cannot be changed."
          schema:
            $ref: "#/definitions/ApiError"
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
    delete:
      tags:
      - "groups"
      summary: "Delete groups"
      description: "Tries to deleteIDMGroup groups with given ids and returns groups\
        \ and statuses of their deletion"
      operationId: "deleteGroups"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Ids of groups to be deleted."
        required: true
        schema:
          type: "array"
          items:
            type: "integer"
            format: "int64"
      responses:
        200:
          description: "Returned HTTP status OK."
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
  /groups/{groupId}:
    get:
      tags:
      - "groups"
      summary: "Get group with given id"
      description: ""
      operationId: "getGroupById"
      produces:
      - "application/json"
      parameters:
      - name: "groupId"
        in: "path"
        description: "Id of group to be returned."
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Group found."
          schema:
            $ref: "#/definitions/GroupDTO"
        404:
          description: "Group cannot be found."
          schema:
            $ref: "#/definitions/ApiError"
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
    delete:
      tags:
      - "groups"
      summary: "Delete group"
      description: "Tries to deleteIDMGroup group with given id and returns if it\
        \ was successful."
      operationId: "deleteGroup"
      produces:
      - "application/json"
      parameters:
      - name: "groupId"
        in: "path"
        description: "Id of group to be deleted."
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Returned deletion status of the group."
        404:
          description: "Group cannot be found."
          schema:
            $ref: "#/definitions/ApiError"
        405:
          description: "Group cannot be deleted because it is a main group."
          schema:
            $ref: "#/definitions/ApiError"
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
  /groups/{groupId}/roles/{roleId}:
    put:
      tags:
      - "groups"
      summary: "Assign role to the group"
      description: ""
      operationId: "assignRoleToGroup"
      parameters:
      - name: "groupId"
        in: "path"
        description: "groupId"
        required: true
        type: "integer"
        format: "int64"
      - name: "roleId"
        in: "path"
        description: "roleId"
        required: true
        type: "integer"
        format: "int64"
      responses:
        204:
          description: "Role assigned to group."
        404:
          description: "Role or group cannot be found."
          schema:
            $ref: "#/definitions/ApiError"
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
    delete:
      tags:
      - "groups"
      summary: "Remove role from the group"
      description: "Role can be removed only if it is not main role of the group."
      operationId: "removeRoleFromGroup"
      parameters:
      - name: "groupId"
        in: "path"
        description: "groupId"
        required: true
        type: "integer"
        format: "int64"
      - name: "roleId"
        in: "path"
        description: "roleId"
        required: true
        type: "integer"
        format: "int64"
      responses:
        204:
          description: "Role successfully removed from the group."
        404:
          description: "Group cannot be found or role cannot be found in group."
          schema:
            $ref: "#/definitions/ApiError"
        409:
          description: "Role cannot be removed from the group because it is the main\
            \ role of the group."
          schema:
            $ref: "#/definitions/ApiError"
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
  /groups/{groupId}/users:
    put:
      tags:
      - "groups"
      summary: "Add users to group."
      description: ""
      operationId: "addUsersToGroup"
      consumes:
      - "application/json"
      parameters:
      - name: "groupId"
        in: "path"
        description: "Id of group to add users."
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "body"
        description: "Ids of members to be added and ids of groups of imported members\
          \ to group."
        required: true
        schema:
          $ref: "#/definitions/AddUsersToGroupDTO"
      responses:
        204:
          description: "User has been given into group."
        304:
          description: "Group is external and cannot be modified."
          schema:
            $ref: "#/definitions/ApiError"
        404:
          description: "Group or some user cannot be found."
          schema:
            $ref: "#/definitions/ApiError"
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
    delete:
      tags:
      - "groups"
      summary: "Remove users from the group."
      description: ""
      operationId: "removeUsers"
      parameters:
      - name: "groupId"
        in: "path"
        description: "Id of group to remove users."
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "body"
        description: "Ids of members to be removed from group."
        required: true
        schema:
          type: "array"
          items:
            type: "integer"
            format: "int64"
      responses:
        204:
          description: "User has been removed from the group."
        304:
          description: "Group is external and cannot be modified."
          schema:
            $ref: "#/definitions/ApiError"
        404:
          description: "Group or some user cannot be found."
          schema:
            $ref: "#/definitions/ApiError"
        409:
          description: "Users cannot be removed from default group or administrator\
            \ cannot remove himself."
          schema:
            $ref: "#/definitions/ApiError"
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
  /groups/{id}/roles:
    get:
      tags:
      - "groups"
      summary: "Returns all roles of group"
      description: ""
      operationId: "getRolesOfGroup"
      parameters:
      - in: "body"
        name: "body"
        description: "Filtering on IDMGroup entity attributes"
        required: false
        schema:
          $ref: "#/definitions/Predicate"
      - name: "id"
        in: "path"
        description: "id"
        required: true
        type: "integer"
        format: "int64"
      - name: "page"
        in: "query"
        description: "Results page you want to retrieve (0..N)"
        required: false
        type: "int"
        x-example: "0"
      - name: "size"
        in: "query"
        description: "Number of records per page."
        required: false
        type: "int"
        x-example: "20"
      - name: "sort"
        in: "query"
        description: "Sorting criteria in the format: property(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        type: "array"
        items:
          type: "string"
          example: "asc"
        collectionFormat: "multi"
        x-example: "asc"
      responses:
        200:
          description: "All roles of group found."
          schema:
            $ref: "#/definitions/RoleRestResource"
        404:
          description: "Group cannot be found."
          schema:
            $ref: "#/definitions/ApiError"
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
...
securityDefinitions:
  bearerAuth:
    description: "Obtain bearer token from OIDC provider and enter the token in the\
      \ following field. Use format 'Bearer {token}'"
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
...
  ApiError:
    type: "object"
    properties:
      timestamp:
        type: "integer"
        format: "int64"
        example: "1574062900 (different for each type of exception)"
        description: "The time when the exception occurred"
      status:
        type: "string"
        example: "404 Not found (different for each type of exception)."
        description: "The HTTP response status code"
        enum:
        - "CONTINUE"
        - "SWITCHING_PROTOCOLS"
        - "PROCESSING"
        - "CHECKPOINT"
        - "OK"
        - "CREATED"
        - "ACCEPTED"
        - "NON_AUTHORITATIVE_INFORMATION"
        - "NO_CONTENT"
        - "RESET_CONTENT"
        - "PARTIAL_CONTENT"
        - "MULTI_STATUS"
        - "ALREADY_REPORTED"
        - "IM_USED"
        - "MULTIPLE_CHOICES"
        - "MOVED_PERMANENTLY"
        - "FOUND"
        - "MOVED_TEMPORARILY"
        - "SEE_OTHER"
        - "NOT_MODIFIED"
        - "USE_PROXY"
        - "TEMPORARY_REDIRECT"
        - "PERMANENT_REDIRECT"
        - "BAD_REQUEST"
        - "UNAUTHORIZED"
        - "PAYMENT_REQUIRED"
        - "FORBIDDEN"
        - "NOT_FOUND"
        - "METHOD_NOT_ALLOWED"
        - "NOT_ACCEPTABLE"
        - "PROXY_AUTHENTICATION_REQUIRED"
        - "REQUEST_TIMEOUT"
        - "CONFLICT"
        - "GONE"
        - "LENGTH_REQUIRED"
        - "PRECONDITION_FAILED"
        - "PAYLOAD_TOO_LARGE"
        - "REQUEST_ENTITY_TOO_LARGE"
        - "URI_TOO_LONG"
        - "REQUEST_URI_TOO_LONG"
        - "UNSUPPORTED_MEDIA_TYPE"
        - "REQUESTED_RANGE_NOT_SATISFIABLE"
        - "EXPECTATION_FAILED"
        - "I_AM_A_TEAPOT"
        - "INSUFFICIENT_SPACE_ON_RESOURCE"
        - "METHOD_FAILURE"
        - "DESTINATION_LOCKED"
        - "UNPROCESSABLE_ENTITY"
        - "LOCKED"
        - "FAILED_DEPENDENCY"
        - "UPGRADE_REQUIRED"
        - "PRECONDITION_REQUIRED"
        - "TOO_MANY_REQUESTS"
        - "REQUEST_HEADER_FIELDS_TOO_LARGE"
        - "UNAVAILABLE_FOR_LEGAL_REASONS"
        - "INTERNAL_SERVER_ERROR"
        - "NOT_IMPLEMENTED"
        - "BAD_GATEWAY"
        - "SERVICE_UNAVAILABLE"
        - "GATEWAY_TIMEOUT"
        - "HTTP_VERSION_NOT_SUPPORTED"
        - "VARIANT_ALSO_NEGOTIATES"
        - "INSUFFICIENT_STORAGE"
        - "LOOP_DETECTED"
        - "BANDWIDTH_LIMIT_EXCEEDED"
        - "NOT_EXTENDED"
        - "NETWORK_AUTHENTICATION_REQUIRED"
      message:
        type: "string"
        example: "The IDMGroup could not be found in database (different for each\
          \ type of exception)."
        description: "The specific description of the ApiError."
      errors:
        type: "array"
        example: "[\"The requested resource was not found (different for each type\
          \ of exception).\"]"
        description: "The list of main reasons of the ApiError."
        items:
          type: "string"
      path:
        type: "string"
        example: "/kypo-rest-user-and-group/api/v1/groups/1000 (different for each\
          \ type of exception)."
        description: "The requested URI path which caused error."
...
  GroupRestResource:
    type: "object"
    properties:
      content:
        type: "array"
        description: "Retrieved IDMGroups from databases."
        items:
          $ref: "#/definitions/GroupDTO"
      pagination:
        description: "Pagination including: page number, number of elements in page,\
          \ size, total elements and total pages."
        $ref: "#/definitions/Pagination"
    description: "Content (Retrieved data) and meta information about REST API result\
      \ page. Including page number, number of elements in page, size of elements,\
      \ total number of elements and total number of pages."
  NewGroupDTO:
    type: "object"
    required:
    - "description"
    - "name"
    properties:
      name:
        type: "string"
        example: "Main group"
        description: "A name of the group."
      description:
        type: "string"
        example: "Group for main users."
        description: "A description of the group."
      expirationDate:
        type: "string"
        format: "date-time"
        example: "2019-11-20T10:28:02.727Z"
        description: "Time until the group is valid."
      users:
        type: "array"
        description: "List of users who is assigned to group."
        uniqueItems: true
        items:
          $ref: "#/definitions/UserForGroupsDTO"
      groupIdsOfImportedUsers:
        type: "array"
        example: "[1]"
        description: "Main identifiers of group."
        items:
          type: "integer"
          format: "int64"
    description: "Data that need to be provided to create a new group."
...
  Predicate:
    type: "object"
...